// src/app/jobseeker/dashboard/components/jobseeker-profile/components/profile-experience/models/addWorkExperience.ts
export class workDetails2 {

  companyName: string;
  jobTitle: string;
  industryType: string;
  currentlyWorking: boolean = false;
  startDate: string;
  endDate: string;
  currentSalary: number;
  salaryType: string;
  description: string;
  location: string;
  empType: string;
  period: string;
  department: string;

}